chore: refactor fixtures, update fixture console, move faker to dev dependencies
