+++ demo/api/block-meta.json
@@ -1754,73 +1754,6 @@
         ]
     },
     {
-        "name": "MediaType",
-        "fields": [
-            {
-                "name": "attachedBlocks",
-                "kind": "NestedObjectList",
-                "object": {
-                    "fields": [
-                        {
-                            "name": "type",
-                            "kind": "String",
-                            "nullable": false
-                        },
-                        {
-                            "name": "props",
-                            "kind": "OneOfBlocks",
-                            "blocks": {
-                                "image": "DamImage",
-                                "damVideo": "DamVideo",
-                                "youTubeVideo": "YouTubeVideo",
-                                "vimeoVideo": "VimeoVideo"
-                            },
-                            "nullable": false
-                        }
-                    ]
-                },
-                "nullable": false
-            },
-            {
-                "name": "activeType",
-                "kind": "String",
-                "nullable": true
-            },
-            {
-                "name": "block",
-                "kind": "NestedObject",
-                "object": {
-                    "fields": [
-                        {
-                            "name": "type",
-                            "kind": "String",
-                            "nullable": false
-                        },
-                        {
-                            "name": "props",
-                            "kind": "OneOfBlocks",
-                            "blocks": {
-                                "image": "DamImage",
-                                "damVideo": "DamVideo",
-                                "youTubeVideo": "YouTubeVideo",
-                                "vimeoVideo": "VimeoVideo"
-                            },
-                            "nullable": false
-                        }
-                    ]
-                },
-                "nullable": true
-            }
-        ],
-        "inputFields": [
-            {
-                "name": "activeType",
-                "kind": "String",
-                "nullable": true
-            }
-        ]
-    },
-    {
         "name": "NewsContent",
         "fields": [
             {
