+++ demo/api/package.json
@@ -29,7 +29,6 @@
     "dependencies": {
         "@comet/blocks-api": "workspace:*",
         "@comet/cms-api": "workspace:*",
-        "@faker-js/faker": "^8.4.1",
         "@hapi/accept": "^5.0.0",
         "@kubernetes/client-node": "^0.18.0",
         "@mikro-orm/cli": "^5.9.8",
@@ -57,6 +56,7 @@
         "base64url": "^3.0.0",
         "class-transformer": "^0.5.0",
         "class-validator": "0.13.2",
+        "cli-progress": "^3.12.0",
         "commander": "^8.0.0",
         "compression": "^1.0.0",
         "cookie-parser": "^1.0.0",
@@ -86,11 +86,12 @@
     },
     "devDependencies": {
         "@comet/eslint-config": "workspace:*",
-        "@faker-js/faker": "^7.6.0",
+        "@faker-js/faker": "8.4.1",
         "@nestjs/cli": "^9.0.0",
         "@nestjs/schematics": "^9.0.0",
         "@nestjs/testing": "^9.0.0",
         "@types/async": "^3.0.0",
+        "@types/cli-progress": "^3.11.6",
         "@types/compression": "^1.0.0",
         "@types/cookie-parser": "^1.0.0",
         "@types/draft-js": "^0.11.10",
