+++ demo/api/src/common/blocks/accordion-item.block.ts
@@ -17,14 +17,17 @@ import { StandaloneCallToActionListBlock } from "@src/common/blocks/standalone-c
 import { StandaloneHeadingBlock } from "@src/common/blocks/standalone-heading.block";
 import { IsBoolean, IsString } from "class-validator";
 
-export const supportedBlocks = {
-    richtext: RichTextBlock,
-    heading: StandaloneHeadingBlock,
-    space: SpaceBlock,
-    callToActionList: StandaloneCallToActionListBlock,
-};
-
-export const AccordionContentBlock = createBlocksBlock({ supportedBlocks }, "AccordionContent");
+export const AccordionContentBlock = createBlocksBlock(
+    {
+        supportedBlocks: {
+            richtext: RichTextBlock,
+            heading: StandaloneHeadingBlock,
+            space: SpaceBlock,
+            callToActionList: StandaloneCallToActionListBlock,
+        },
+    },
+    "AccordionContent",
+);
 
 class AccordionItemBlockData extends BlockData {
     @BlockField({ nullable: true })
