+++ demo/api/src/db/fixtures/fixtures.module.ts
@@ -13,21 +13,15 @@ import { Product } from "@src/products/entities/product.entity";
 import { ConsoleModule } from "nestjs-console";
 
 import { AccordionBlockFixtureService } from "./generators/blocks/layout/accordion-block-fixture.service";
-import { AccordionContentBlockFixtureService } from "./generators/blocks/layout/accordion-content-block-fixture.service";
-import { AccordionItemBlockFixtureService } from "./generators/blocks/layout/accordion-item-block-fixture.service";
 import { ColumnsBlockFixtureService } from "./generators/blocks/layout/columns-block-fixture.service";
-import { ColumnsContentBlockFixtureService } from "./generators/blocks/layout/columns-content-block-fixture.service";
 import { ContentGroupBlockFixtureService } from "./generators/blocks/layout/content-group-block-fixture.service";
-import { ContentGroupContentBlockFixtureService } from "./generators/blocks/layout/content-group-content-block-fixture.service";
 import { LayoutBlockFixtureService } from "./generators/blocks/layout/layout-block-fixture.service";
 import { SpaceBlockFixtureService } from "./generators/blocks/layout/space-block-fixture.service";
 import { DamImageBlockFixtureService } from "./generators/blocks/media/dam-image-block-fixture.service";
 import { DamVideoBlockFixtureService } from "./generators/blocks/media/dam-video-block-fixture.service";
 import { FullWidthImageBlockFixtureService } from "./generators/blocks/media/full-width-image-block-fixture.service";
+import { MediaBlockFixtureService } from "./generators/blocks/media/media-block.fixture.service";
 import { MediaGalleryBlockFixtureService } from "./generators/blocks/media/media-gallery-block-fixture.service";
-import { MediaGalleryItemBlockFixtureService } from "./generators/blocks/media/media-gallery-item-block-fixture.service";
-import { MediaGalleryListBlockFixtureService } from "./generators/blocks/media/media-gallery-list-block-fixture.service";
-import { MediaTypeBlockFixtureService } from "./generators/blocks/media/media-type-block.fixture.service";
 import { PixelImageBlockFixtureService } from "./generators/blocks/media/pixel-image-block-fixture.service";
 import { StandaloneMediaBlockFixtureService } from "./generators/blocks/media/standalone-media-block-fixture.service";
 import { SvgImageBlockFixtureService } from "./generators/blocks/media/svg-image-block-fixture.service";
@@ -43,10 +37,8 @@ import { TextLinkBlockFixtureService } from "./generators/blocks/navigation/text
 import { BasicStageBlockFixtureService } from "./generators/blocks/stage/basic-stage-block-fixture.service";
 import { BillboardTeaserBlockFixtureService } from "./generators/blocks/teaser/billboard-teaser-block-fixture.service";
 import { TeaserBlockFixtureService } from "./generators/blocks/teaser/teaser-block-fixture.service";
-import { TeaserItemBlockFixtureService } from "./generators/blocks/teaser/teaser-item-block-fixture.service";
 import { HeadingBlockFixtureService } from "./generators/blocks/text-and-content/heading-block-fixture.service";
 import { KeyFactsBlockFixtureService } from "./generators/blocks/text-and-content/key-facts-block-fixture.service";
-import { KeyFactsItemBlockFixtureService } from "./generators/blocks/text-and-content/key-facts-item-block-fixture.service";
 import { RichTextBlockFixtureService } from "./generators/blocks/text-and-content/rich-text-block-fixture.service";
 import { StandaloneHeadingBlockFixtureService } from "./generators/blocks/text-and-content/standalone-heading-block-fixture.service";
 import { TextImageBlockFixtureService } from "./generators/blocks/text-and-content/text-image-block-fixture.service";
@@ -73,8 +65,6 @@ import { VideoFixtureService } from "./generators/video-fixture.service";
         MikroOrmModule.forFeature([DamFile, Page, Link, Product, Manufacturer]),
     ],
     providers: [
-        AccordionContentBlockFixtureService,
-        AccordionItemBlockFixtureService,
         AccordionBlockFixtureService,
         AnchorBlockFixtureService,
         BasicStageBlockFixtureService,
@@ -82,9 +72,7 @@ import { VideoFixtureService } from "./generators/video-fixture.service";
         CallToActionBlockFixtureService,
         CallToActionListBlockFixtureService,
         ColumnsBlockFixtureService,
-        ColumnsContentBlockFixtureService,
         ContentGroupBlockFixtureService,
-        ContentGroupContentBlockFixtureService,
         DamImageBlockFixtureService,
         DamVideoBlockFixtureService,
         DocumentGeneratorService,
@@ -95,15 +83,12 @@ import { VideoFixtureService } from "./generators/video-fixture.service";
         ImageFileFixtureService,
         ImageFixtureService,
         KeyFactsBlockFixtureService,
-        KeyFactsItemBlockFixtureService,
         LayoutBlockFixtureService,
         LinkBlockFixtureService,
         LinkListBlockFixtureService,
         ManyImagesTestPageFixtureService,
         MediaGalleryBlockFixtureService,
-        MediaGalleryItemBlockFixtureService,
-        MediaGalleryListBlockFixtureService,
-        MediaTypeBlockFixtureService,
+        MediaBlockFixtureService,
         PageContentBlockFixtureService,
         PixelImageBlockFixtureService,
         RedirectsFixtureService,
@@ -119,7 +104,6 @@ import { VideoFixtureService } from "./generators/video-fixture.service";
         StandaloneHeadingBlockFixtureService,
         StandaloneMediaBlockFixtureService,
         TeaserBlockFixtureService,
-        TeaserItemBlockFixtureService,
         TextImageBlockFixtureService,
         TextLinkBlockFixtureService,
         VideoFixtureService,
