+++ demo/api/src/db/fixtures/generators/blocks/layout/layout-block-fixture.service.ts
@@ -3,22 +3,22 @@ import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { LayoutBlock, LayoutBlockLayout } from "@src/documents/pages/blocks/layout.block";
 
-import { MediaTypeBlockFixtureService } from "../media/media-type-block.fixture.service";
+import { MediaBlockFixtureService } from "../media/media-block.fixture.service";
 import { RichTextBlockFixtureService } from "../text-and-content/rich-text-block-fixture.service";
 
 @Injectable()
 export class LayoutBlockFixtureService {
     constructor(
-        private readonly mediaTypeBlockFixtureService: MediaTypeBlockFixtureService,
+        private readonly mediaBlockFixtureService: MediaBlockFixtureService,
         private readonly richTextFixtureService: RichTextBlockFixtureService,
     ) {}
 
     async generateBlockInput(): Promise<ExtractBlockInputFactoryProps<typeof LayoutBlock>> {
         return {
             layout: faker.helpers.arrayElement(Object.values(LayoutBlockLayout)),
-            media1: await this.mediaTypeBlockFixtureService.generateBlockInput(),
+            media1: await this.mediaBlockFixtureService.generateBlockInput(),
             text1: await this.richTextFixtureService.generateBlockInput(),
-            media2: await this.mediaTypeBlockFixtureService.generateBlockInput(),
+            media2: await this.mediaBlockFixtureService.generateBlockInput(),
             text2: await this.richTextFixtureService.generateBlockInput(),
         };
     }
