+++ demo/api/src/db/fixtures/generators/blocks/teaser/billboard-teaser-block-fixture.service.ts
@@ -3,7 +3,7 @@ import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { BillboardTeaserBlock } from "@src/documents/pages/blocks/billboard-teaser.block";
 
-import { MediaTypeBlockFixtureService } from "../media/media-type-block.fixture.service";
+import { MediaBlockFixtureService } from "../media/media-block.fixture.service";
 import { CallToActionListBlockFixtureService } from "../navigation/call-to-action-list-block.service";
 import { HeadingBlockFixtureService } from "../text-and-content/heading-block-fixture.service";
 import { RichTextBlockFixtureService } from "../text-and-content/rich-text-block-fixture.service";
@@ -13,13 +13,13 @@ export class BillboardTeaserBlockFixtureService {
     constructor(
         private readonly callToActionListBlockFixtureService: CallToActionListBlockFixtureService,
         private readonly headingBlockFixtureService: HeadingBlockFixtureService,
-        private readonly mediaTypeBlockFixtureService: MediaTypeBlockFixtureService,
+        private readonly mediaBlockFixtureService: MediaBlockFixtureService,
         private readonly richTextBlockFixtureService: RichTextBlockFixtureService,
     ) {}
 
     async generateBlockInput(): Promise<ExtractBlockInputFactoryProps<typeof BillboardTeaserBlock>> {
         return {
-            media: await this.mediaTypeBlockFixtureService.generateBlockInput(),
+            media: await this.mediaBlockFixtureService.generateBlockInput(),
             heading: await this.headingBlockFixtureService.generateBlockInput(),
             text: await this.richTextBlockFixtureService.generateBlockInput(),
             overlay: faker.number.int({ min: 50, max: 90 }),
