+++ demo/api/src/db/fixtures/generators/blocks/teaser/teaser-block-fixture.service.ts
@@ -2,12 +2,28 @@ import { ExtractBlockInputFactoryProps } from "@comet/blocks-api";
 import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { TeaserBlock } from "@src/documents/pages/blocks/teaser.block";
+import { TeaserItemBlock } from "@src/documents/pages/blocks/teaser-item.block";
 
-import { TeaserItemBlockFixtureService } from "./teaser-item-block-fixture.service";
+import { MediaBlockFixtureService } from "../media/media-block.fixture.service";
+import { TextLinkBlockFixtureService } from "../navigation/text-link-block-fixture.service";
+import { RichTextBlockFixtureService } from "../text-and-content/rich-text-block-fixture.service";
 
 @Injectable()
 export class TeaserBlockFixtureService {
-    constructor(private readonly teaserItemBlockFixtureService: TeaserItemBlockFixtureService) {}
+    constructor(
+        private readonly mediaBlockFixtureService: MediaBlockFixtureService,
+        private readonly richTextBlockFixtureService: RichTextBlockFixtureService,
+        private readonly textLinkBlockFixtureService: TextLinkBlockFixtureService,
+    ) {}
+
+    async generateTeaserItemBlock(): Promise<ExtractBlockInputFactoryProps<typeof TeaserItemBlock>> {
+        return {
+            media: await this.mediaBlockFixtureService.generateBlockInput(),
+            title: faker.lorem.words({ min: 3, max: 9 }),
+            description: await this.richTextBlockFixtureService.generateBlockInput(),
+            link: await this.textLinkBlockFixtureService.generateBlockInput(),
+        };
+    }
 
     async generateBlockInput(min = 2, max = 6): Promise<ExtractBlockInputFactoryProps<typeof TeaserBlock>> {
         const blockAmount = faker.number.int({ min, max });
@@ -17,7 +33,7 @@ export class TeaserBlockFixtureService {
             blocks.push({
                 key: faker.string.uuid(),
                 visible: true,
-                props: await this.teaserItemBlockFixtureService.generateBlockInput(),
+                props: await this.generateTeaserItemBlock(),
             });
         }
