+++ demo/api/src/db/fixtures/generators/blocks/text-and-content/key-facts-block-fixture.service.ts
@@ -2,12 +2,26 @@ import { ExtractBlockInputFactoryProps } from "@comet/blocks-api";
 import { faker } from "@faker-js/faker";
 import { Injectable } from "@nestjs/common";
 import { KeyFactsBlock } from "@src/documents/pages/blocks/key-facts.block";
+import { KeyFactsItemBlock } from "@src/documents/pages/blocks/key-facts-item.block";
 
-import { KeyFactsItemBlockFixtureService } from "./key-facts-item-block-fixture.service";
+import { SvgImageBlockFixtureService } from "../media/svg-image-block-fixture.service";
+import { RichTextBlockFixtureService } from "./rich-text-block-fixture.service";
 
 @Injectable()
 export class KeyFactsBlockFixtureService {
-    constructor(private readonly keyFactsItemBlockFixtureService: KeyFactsItemBlockFixtureService) {}
+    constructor(
+        private readonly svgImageBlockFixtureService: SvgImageBlockFixtureService,
+        private readonly richTextBlockFixtureService: RichTextBlockFixtureService,
+    ) {}
+
+    async generateKeyFactsItemBlock(): Promise<ExtractBlockInputFactoryProps<typeof KeyFactsItemBlock>> {
+        return {
+            icon: await this.svgImageBlockFixtureService.generateBlockInput(),
+            fact: faker.lorem.sentence(),
+            label: faker.lorem.sentence(),
+            description: await this.richTextBlockFixtureService.generateBlockInput(),
+        };
+    }
 
     async generateBlockInput(min = 2, max = 6): Promise<ExtractBlockInputFactoryProps<typeof KeyFactsBlock>> {
         const blockAmount = faker.number.int({ min, max });
@@ -17,7 +31,7 @@ export class KeyFactsBlockFixtureService {
             blocks.push({
                 key: faker.string.uuid(),
                 visible: true,
-                props: await this.keyFactsItemBlockFixtureService.generateBlockInput(),
+                props: await this.generateKeyFactsItemBlock(),
             });
         }
