+++ demo/api/src/documents/pages/blocks/content-group.block.ts
@@ -23,21 +23,24 @@ import { KeyFactsBlock } from "@src/documents/pages/blocks/key-facts.block";
 import { TeaserBlock } from "@src/documents/pages/blocks/teaser.block";
 import { IsEnum } from "class-validator";
 
-export const supportedBlocks = {
-    accordion: AccordionBlock,
-    anchor: AnchorBlock,
-    callToActionList: StandaloneCallToActionListBlock,
-    columns: ColumnsBlock,
-    heading: StandaloneHeadingBlock,
-    keyFacts: KeyFactsBlock,
-    media: StandaloneMediaBlock,
-    mediaGallery: MediaGalleryBlock,
-    richtext: RichTextBlock,
-    space: SpaceBlock,
-    teaser: TeaserBlock,
-};
-
-export const ContentBlock = createBlocksBlock({ supportedBlocks }, { name: "ContentGroupContent" });
+export const ContentBlock = createBlocksBlock(
+    {
+        supportedBlocks: {
+            accordion: AccordionBlock,
+            anchor: AnchorBlock,
+            callToActionList: StandaloneCallToActionListBlock,
+            columns: ColumnsBlock,
+            heading: StandaloneHeadingBlock,
+            keyFacts: KeyFactsBlock,
+            media: StandaloneMediaBlock,
+            mediaGallery: MediaGalleryBlock,
+            richtext: RichTextBlock,
+            space: SpaceBlock,
+            teaser: TeaserBlock,
+        },
+    },
+    { name: "ContentGroupContent" },
+);
 
 export enum BackgroundColor {
     default = "default",
